requirements.txt
langchain
langchain-google-genai
langgraph
python-dotenv
twilio
fastapi
uvicorn
langchain-chroma
langchain-text-splitters
langchain-community
pypdf
fpdf2
aiosqlite
tenacity
